Create gui-iracing.h
# gui-iracing.cxx

This module is a part of the Visor application, specifically handling the iRacing telemetry GUI.

## Functions

### startup()
This function serves as an initializer for the iRacing GUI module. Currently, it doesn't perform any operations but stands as a placeholder for potential initialization tasks.

### shutdown()
This function serves as a deinitializer for the iRacing GUI module. Similarly to `startup()`, it currently doesn't perform any operations but stands as a placeholder for potential cleanup tasks.

### emit_content()
This function is responsible for rendering or updating the iRacing telemetry GUI content.

The GUI consists of a tab bar labeled "iRacingTabBar" with a single tab item named "Telemetry". Inside this tab, the function displays various iRacing telemetry variables such as:

- Connection status represented by different color-coded texts (Connected, Live, Searching, and Paused)
- Lap percentage, represented by both a progress bar and text
- RPM (Revolutions per Minute)
- Speed
- Current gear

The function also loops through all other telemetry variables available through the iRacing API, displaying each variable name and type in a child window with the identifier "TelemetryVariables".