Add username input to account creation process
Modified the “Create Account” feature to include a new input field for username during the account creation process. Added ImGui::InputTextWithHint() function to display an input box for the username field. Updated the api::customer::create_new() function call to include the username parameter with the value from the input field. This enhancement allows users to create an account with a desired username in addition to their name, email, and password.